backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
torchaudio==2.4.0
torch==2.4.0
librosa==0.10.2.post1
opensmile==2.5.0
praat-parselmouth==0.4.3
transformers==4.45.2
scikit-learn==1.5.2
xgboost==2.1.1
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2
soundfile==0.12.1
python-dotenv==1.0.1
